The concept of visual future: smart glasses using Raspberry Pi
The main aim of this project is to place computers’ and smartphones’ content in the user’s field of view or ears. The content includes augmented reality, talking system, heads-up displays, and secure connection. Instead of looking at a rectangular screen, we will see words, pictures, objects and virtual environments by simply looking around. Many of the devices which is available in the market have almost the same features like access message, calls, tracking locations and provide GPS maps navigation. The project came up with different features that do not exist at least in the market yet, but few people who have been working with it, because it is difficult to find the software sources that can support this kind of project. The main features include image detection, American sign languages and voice assistant.